backend/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
h11==0.14.0
idna==3.10
sniffio==1.3.1
typing-extensions==4.12.2
urllib3==2.2.3
annotated-types==0.7.0
anyio==4.6.2.post1
colorama==0.4.6
httpcore==1.0.6
protobuf==4.25.5
pyasn1==0.6.1
pydantic-core==2.23.4
requests==2.32.3
cachetools==5.5.0
filelock==3.16.1
fsspec==2024.10.0
googleapis-common-protos==1.65.0
grpcio==1.67.1
httpx==0.27.2
jsonpointer==3.0.0
orjson==3.10.11
packaging==24.1
pyasn1-modules==0.4.1
pydantic==2.9.2
pyparsing==3.2.0
pyyaml==6.0.2
requests-toolbelt==1.0.0
rsa==4.9
tqdm==4.66.6
frozenlist==1.5.0
google-auth==2.35.0
grpcio-status==1.62.3
httplib2==0.22.0
huggingface-hub==0.26.2
jsonpatch==1.33
langsmith==0.1.139
multidict==6.1.0
propcache==0.2.0
proto-plus==1.25.0
tenacity==9.0.0
aiohappyeyeballs==2.4.3
aiosignal==1.3.1
attrs==24.2.0
distro==1.9.0
google-api-core==2.22.0
google-auth-httplib2==0.2.0
greenlet==3.1.1
httpx-sse==0.4.0
jiter==0.7.0
langchain-core==0.3.15
msgpack==1.1.0
mypy-extensions==1.0.0
regex==2024.9.11
tokenizers==0.20.1
uritemplate==4.1.1
yarl==1.17.1
aiohttp==3.10.10
anthropic==0.38.0
defusedxml==0.7.1
google-ai-generativelanguage==0.6.6
google-api-python-client==2.151.0
langchain-text-splitters==0.3.2
langgraph-checkpoint==2.0.2
langgraph-sdk==0.1.35
marshmallow==3.23.1
numpy==1.26.4
openai==1.53.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
starlette==0.41.2
tiktoken==0.8.0
typing-inspect==0.9.0
click==8.1.7
dataclasses-json==0.6.7
fastapi==0.115.4
google-generativeai==0.7.2
langchain==0.3.4
langchain-anthropic==0.2.3
langchain-openai==0.2.3
langgraph==0.2.44
partialjson==0.0.8
pydantic-settings==2.6.1
toml==0.10.2
copilotkit==0.1.27
html2text==2024.2.26
langchain-community==0.3.3
langchain-google-genai==2.0.0
tavily-python==0.5.0
uvicorn==0.31.1
langchain-pinecone
snowflake-connector-python
markdown
weasyprint